fix: facility control admin view
